requirements.txt
flask>=2.0.0
langchain-google-genai
langchain-chroma
langchain-community
python-dotenv
tqdm